% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcels.R
\name{address.update}
\alias{address.update}
\title{address.update}
\usage{
address.update(pkg.data.root, parcels.address, parcels.address.geocode.new,
  parcel.source = "points")
}
\arguments{
\item{pkg.data.root}{dropbox package root}

\item{parcels.address}{data.table}

\item{parcels.address.geocode.new}{geocoded info spatially intersected}

\item{parcel.source}{where it conmes from (points)}
}
\description{
Updates address file with intersected coordinates
}
\keyword{clean}
\keyword{parcels,}
